requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
numpy
scikit-learn
joblib
python-multipart
sentence-transformers
faiss-cpu
openai
transformers
torch
tensorflow
lifelines
shap
matplotlib
plotly
